Remove validators file
